+++ RELEASE_NOTES.md
@@ -81,3 +81,8 @@
 
 #### 0.0.18 - 
 * Add GetAllUsesOfAllSymbols and GetAllUsesOfAllSymbolsInFile
+
+#### 0.0.18 - 
+* Change return type of GetAllUsesOfSymbol, GetAllUsesOfAllSymbols and GetAllUsesOfAllSymbolsInFile to FSharpSymbolUse
+* Add symbol uses when an abstract member is implemented.
+